New code base for CentOS 7 on VCL.

+++ sbin/cluster.py
@@ -34,9 +34,11 @@ def cmd(ctx, cmdline):
 
 @cli.command()
 @click.option('-p', '--prefix', default='node')
+@click.option('--overwrite', is_flag=True)
+@click.option('--os', type=click.Choice(['ubuntu', 'centos']), default='ubuntu')
 @click.pass_context
-def create_hosts(ctx, prefix):
-    preload_config = '''127.0.0.1 localhost
+def create_hosts(ctx, prefix, overwrite, os):
+    ubuntu_preload_config = '''127.0.0.1 localhost
 
 # The following lines are desirable for IPv6 capable hosts
 ::1 ip6-localhost ip6-loopback
@@ -46,14 +48,28 @@ ff02::1 ip6-allnodes
 ff02::2 ip6-allrouters
 ff02::3 ip6-allhosts'''
 
+    centos_preload_config = '''127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
+::1         localhost localhost.localdomain localhost6 localhost6.localdomain6'''
+
     tmp_file = '/tmp/.hosts'
 
-    with open(tmp_file, 'w') as f:
-        f.write(preload_config)
-        f.write('\n\n')
+    if overwrite:
+        with open(tmp_file, 'w') as f:
+            if os == 'cento':
+                f.write(centos_preload_config)
+            else:
+                f.write(ubuntu_preload_config)
+            f.write('\n\n')
+            for i in range(len(ctx.obj['host_list'])):
+                host = ctx.obj['host_list'][i]
+                f.write("{} {}{}\n".format(host, prefix, i+1))
+    else:
+        execute('cp /etc/hosts {}'.format(tmp_file))
+        execute('echo >> {}'.format(tmp_file))
         for i in range(len(ctx.obj['host_list'])):
             host = ctx.obj['host_list'][i]
-            f.write("{} {}{}\n".format(host, prefix, i+1))
+            host_mapping = "{} {}{}".format(host, prefix, i + 1)
+            execute('echo {} >> {}'.format(host_mapping, tmp_file))
 
     with parallel.CommandAgent(show_result=False) as agent:
         for host in ctx.obj['host_list']:
